modified bla bla bla

+++ Employee.java
@@ -1,11 +1,11 @@
 public class Employee extends Person {
 
     private Role role;
-    private float salary;
+    private double salary;
     private Company company;
 
 
-    public Employee(int id, String firstName, String lastName, Role role, Float salary, Company company) {
+    public Employee(int id, String firstName, String lastName, Role role, double salary, Company company) {
         setId(id);
         setFirstName(firstName);
         setLastName(lastName);
@@ -22,11 +22,11 @@ public class Employee extends Person {
         this.role = role;
     }
 
-    public float getSalary() {
+    public double getSalary() {
         return salary;
     }
 
-    public void setSalary(float salary) {
+    public void setSalary(double salary) {
         this.salary = salary;
     }
